Move web3 method v2 files to _utils folder
